Group all homeassistant/core into a module
