+++ supervisor/homeassistant/api.py
@@ -1,0 +1,124 @@
+"""Home Assistant control object."""
+import asyncio
+from contextlib import asynccontextmanager, suppress
+from datetime import datetime, timedelta
+import logging
+from typing import Any, AsyncContextManager, Dict, Optional
+
+import aiohttp
+from aiohttp import hdrs
+
+from ..coresys import CoreSys, CoreSysAttributes
+from ..exceptions import HomeAssistantAPIError, HomeAssistantAuthError
+from ..utils import check_port
+
+_LOGGER: logging.Logger = logging.getLogger(__name__)
+
+
+class HomeAssistantAPI(CoreSysAttributes):
+    """Home Assistant core object for handle it."""
+
+    def __init__(self, coresys: CoreSys):
+        """Initialize Home Assistant object."""
+        self.coresys: CoreSys = coresys
+
+        # We don't persist access tokens. Instead we fetch new ones when needed
+        self._access_token: Optional[str] = None
+        self._access_token_expires: Optional[datetime] = None
+
+    async def ensure_access_token(self) -> None:
+        """Ensure there is an access token."""
+        if (
+            self._access_token is not None
+            and self._access_token_expires > datetime.utcnow()
+        ):
+            return
+
+        with suppress(asyncio.TimeoutError, aiohttp.ClientError):
+            async with self.sys_websession_ssl.post(
+                f"{self.api_url}/auth/token",
+                timeout=30,
+                data={
+                    "grant_type": "refresh_token",
+                    "refresh_token": self.refresh_token,
+                },
+            ) as resp:
+                if resp.status != 200:
+                    _LOGGER.error("Can't update Home Assistant access token!")
+                    raise HomeAssistantAuthError()
+
+                _LOGGER.info("Updated Home Assistant API token")
+                tokens = await resp.json()
+                self._access_token = tokens["access_token"]
+                self._access_token_expires = datetime.utcnow() + timedelta(
+                    seconds=tokens["expires_in"]
+                )
+
+    @asynccontextmanager
+    async def make_request(
+        self,
+        method: str,
+        path: str,
+        json: Optional[Dict[str, Any]] = None,
+        content_type: Optional[str] = None,
+        data: Any = None,
+        timeout: int = 30,
+        params: Optional[Dict[str, str]] = None,
+        headers: Optional[Dict[str, str]] = None,
+    ) -> AsyncContextManager[aiohttp.ClientResponse]:
+        """Async context manager to make a request with right auth."""
+        url = f"{self.sys_homeassistant.api_url}/{path}"
+        headers = headers or {}
+
+        # Passthrough content type
+        if content_type is not None:
+            headers[hdrs.CONTENT_TYPE] = content_type
+
+        for _ in (1, 2):
+            # Prepare Access token
+            if self.refresh_token:
+                await self.ensure_access_token()
+                headers[hdrs.AUTHORIZATION] = f"Bearer {self._access_token}"
+
+            try:
+                async with getattr(self.sys_websession_ssl, method)(
+                    url,
+                    data=data,
+                    timeout=timeout,
+                    json=json,
+                    headers=headers,
+                    params=params,
+                ) as resp:
+                    # Access token expired
+                    if resp.status == 401 and self.sys_homeassistant.refresh_token:
+                        self._access_token = None
+                        continue
+                    yield resp
+                    return
+            except (asyncio.TimeoutError, aiohttp.ClientError) as err:
+                _LOGGER.error("Error on call %s: %s", url, err)
+                break
+
+        raise HomeAssistantAPIError()
+
+    async def check_api_state(self) -> bool:
+        """Return True if Home Assistant up and running."""
+        # Check if port is up
+        if not await self.sys_run_in_executor(
+            check_port,
+            self.sys_homeassistant.ip_address,
+            self.sys_homeassistant.api_port,
+        ):
+            return False
+
+        # Check if API is up
+        with suppress(HomeAssistantAPIError):
+            async with self.make_request("get", "api/config") as resp:
+                if resp.status in (200, 201):
+                    data = await resp.json()
+                    if data.get("state", "RUNNING") == "RUNNING":
+                        return True
+                else:
+                    _LOGGER.debug("Home Assistant API return: %d", resp.status)
+
+        return False
